\name{CD}
\alias{CD-class}
\alias{CD}
\alias{show,CD-method}
\alias{dat2save,CD-method}
\alias{explo.plot,CD-method}
\docType{class}
\title{CD class}

\description{
CD class generated from dat() function with type="cd". This object contains the cumulative distributions of reads for the selected samples.
}

\usage{
\S4method{explo.plot}{CD}(object)
\S4method{dat2save}{CD}(object)
}

\arguments{
  \item{object}{
    Object generated from \code{dat()} function.
  }
}

\section{Slots/List Components}{
Objects of this class contain (at least) the following list components:

  \code{dat}: Data.frame containing the information generated by dat() function: percentage of detected features and
              cumulative percentage of reads associated to these features in selected samples.
  
}

\section{Methods}{
This class has an specific \code{show} method in order to work and print
a summary of the elements which are contained and a \code{dat2save} method
to save the relevant information in an object cleanly. It also has an
\code{explo.plot} method to plot the data contained in the object.
}

\author{Sonia Tarazona}

\keyword{classes}
