\name{DLBio}
\alias{DLBio-class}
\alias{DLBio}
\alias{explo.plot,DLBio-method}
\alias{show,DLBio-method}
\alias{dat2save,DLBio-method}
\docType{class}
\title{DLBio class}

\description{
DLBio class generated from dat() function with type="DLbio". This object contains the median length of the detected features for each biotype
at increasing sequencing depths.
}

\usage{
\S4method{explo.plot}{DLBio}(object, samples = NULL, toplot = "protein_coding",ylim = NULL)
\S4method{dat2save}{DLBio}(object)
}

\arguments{
  \item{object}{
    Object generated from \code{dat()} function.
  }
  \item{toplot}{
    This parameter indicates which biological group is to be plotted.
    It may be a number or a text with the name of the biological group. 
  }
  \item{samples}{
    The samples to be plotted. If NULL, the two first samples are plotted.
  }
  \item{ylim}{
    Range for Y axis. If NULL (default), an appropriate range is computed.
  }
}

\section{Slots/List Components}{
Objects of this class contain (at least) the following list components:

  \code{dat}: List containing the information generated by dat() function. This list has the following elements:
  
  \code{result}: A list with as many elements as biological groups. Each biological group is another list with all the selected samples, 
		and for each sample, the median length of the detected features at each sequencing depth. 

  \code{bionum}: A list with as many elements as biological groups, and for each one, the number of features in that biological group.
  
  \code{depth}: A list with as many elements as selected samples. For each one, the increasing sequencing depths to be plotted.
		The last sequencing depth is the real one and the others are simulated from it using the multinomial distribution.

  \code{biolength}: A list with as many elements as biological groups. For each one, the median length of the features in that group.
  
}

\section{Methods}{
This class has an specific \code{show} method in order to work and print
a summary of the elements which are contained and a \code{dat2save} method
to save the relevant information in an object cleanly. It also has an
\code{explo.plot} method to plot the data contained in the object.
}

\author{Sonia Tarazona}

\keyword{classes}
