\name{example}

\docType{data}

\alias{mydata}
\alias{mynoiseq}

\title{Example of objects used and created by the NOISeq package}
\description{
  This is a quick view of the objects generated by the
  package. To take a look, see the usage information. These objects have
  been created from Marioni's reduce dataset (only chromosomes I to IV).
}
\usage{
# To load the object myCounts generated by the readData() function from R objects containing expression data, the factors describing the experimental conditions to be studied, the feature length, the feature biotypes, the chromosome and the position:
data(myCounts)

# To load the object generated after running the noiseq() function to compute differential expression:
data(noiseq)
}

\references{
    Marioni, J.C. and Mason, C.E. and Mane, S.M. and Stephens, M. and Gilad, Y. (2008) RNA-seq: an assessment of
technical reproducibility and comparison with gene expression arrays. \emph{Genome Research}, \bold{18}:
1509--1517.
}

\keyword{datasets}
