\name{Saturation}
\alias{Saturation-class}
\alias{Saturation}
\alias{explo.plot,Saturation-method}
\alias{show,Saturation-method}
\alias{dat2save,Saturation-method}
\docType{class}
\title{Saturation class}

\description{
Saturation class generated from dat() function with type="biodetection". This object contains the number of detected features per biotype at 
increasing sequencing depths and also the new detections when increasing the sequencing depth in one million of reads.
}

\usage{
\S4method{explo.plot}{Saturation}(object, toplot = 1, samples = NULL,ylim = NULL, yrightlim = NULL)
\S4method{dat2save}{Saturation}(object)
}

\arguments{
  \item{object}{
    Object generated from \code{dat()} function.
  }
   \item{toplot}{
    This parameter indicates which biological group is to be plotted.
    It may be a number or a text with the name of the biological group. 
    If toplot=1 (or "global"), a global plot with all the biological groups will be generated.
  }
  \item{samples}{
    The samples to be plotted. If NULL, all the samples are plotted for Saturation object.
  }
  \item{ylim}{
    Range for Y axis (on the left-hand side of the plot). If NULL (default), an appropriate range is computed.
  }
  \item{yrightlim}{
    Range for Y right-axis (on the right-hand side of the plot). If NULL (default), an appropriate range is computed.    
    }
     
}

\section{Slots/List Components}{
Objects of this class contain (at least) the following list components:

  \code{dat}: List containing the information generated by dat() function. This list has the following elements:
  
  \code{saturation}: List containing for all the biological classes (and also a global class with all of them together) the saturation data 
		     to be plotted for each selected sample (in Y left axis).

  \code{bionum}: List containing for all the biological classes (and also a global class with all of them together) the number of 
		 features for that group.
  
  \code{depth}: List containing for each selected sample the increasing sequencing depths to be plotted.

  \code{newdet}: List containing for all the biological classes (and also a global class with all of them together) the new detection data 
		 to be plotted for each selected sample (in Y right axis).
  
}

\section{Methods}{
This class has an specific \code{show} method in order to work and print
a summary of the elements which are contained and a \code{dat2save} method
to save the relevant information in an object cleanly. It also has an
\code{explo.plot} method to plot the data contained in the object.
}

\author{Sonia Tarazona}

\keyword{classes}
