\name{Biodetection}
\alias{Biodetection-class}
\alias{Biodetection}
\alias{show,Biodetection-method}
\alias{explo.plot,Biodetection-method}
\alias{dat2save,Biodetection-method}
\docType{class}
\title{Biodetection class}

\description{
Biodetection class generated from dat() function with type="biodetection". This object contains the percentage of each biological 
class (e.g. biotype) in the genome (i.e. in the whole set of features provided), the corresponding percentage detected by the sample 
and the percentage of the biotype within the sample.
}

\usage{
\S4method{explo.plot}{Biodetection}(object)
\S4method{dat2save}{Biodetection}(object)
}

\arguments{
  \item{object}{
    Object generated from \code{dat()} function.
  }
}

\section{Slots/List Components}{
Objects of this class contain (at least) the following list components:

  \code{dat}: List containing the information generated by dat() function. This list has the following elements:
  
  \code{table,table2}: Table(s) containing the results for the selected sample(s).

  \code{params}: Maximum values for the Y-axes scale in the plot generated from these data (explo.plot).
  
  \code{samples}: Names of the sample(s) evaluated by dat(). 
  
}

\section{Methods}{
This class has an specific \code{show} method in order to work and print
a summary of the elements which are contained and a \code{dat2save} method
to save the relevant information in an object cleanly. It also has an
\code{explo.plot} method to plot the data contained in the object.
}

\author{Sonia Tarazona}

\keyword{classes}
