\name{CountsBio}
\alias{CountsBio-class}
\alias{CountsBio}
\alias{show,CountsBio-method}
\alias{explo.plot,CountsBio-method}
\alias{dat2save,CountsBio-method}
\docType{class}
\title{CountsBio class}

\description{
CountsBio class generated from dat() function with
type="countsbio". This object contains the count distribution for each biological group and also for increasing sequencing depths.
}

\usage{
\S4method{explo.plot}{CountsBio}(object, toplot = 1, samples = NULL, ylim = NULL)
\S4method{dat2save}{CountsBio}(object)
}

\arguments{
  \item{object}{
    Object generated from \code{dat()} function.
  }
  \item{toplot}{
    This parameter indicates which biological group is to be plotted.
    It may be a number or a text with the name of the biological group. 
    If toplot=1 (or "global"), a global plot with all the biological groups will be generated.

  }
  \item{samples}{
    The samples to be plotted. If NULL, the two first samples are plotted because the plot for this object only admit a maximum of two samples.
  }
  \item{ylim}{
    Range for the Y axis. If NULL (default), an appropriate range is computed.
  }
}

\section{Slots/List Components}{
Objects of this class contain (at least) the following list components:

  \code{dat}: List containing the information generated by dat() function. This list has the following elements:
  
  \code{result}: List containing for all the biological classes (and also a global class with all of them together) the data to be plotted for
		 each selected sample.

  \code{bionum}: List containing for all the biological classes (and also a global class with all of them together) the number of 
		 features for that group.
  
  \code{depth}: List containing for each selected sample the increasing sequencing depths to be plotted.

  \code{quart}: List containing for each selected sample a matrix summarizing the quartiles of the distributions of counts for each biotype at that total sequencing depth in each
		selected sample.
  
}

\section{Methods}{
This class has an specific \code{show} method in order to work and print
a summary of the elements which are contained and a \code{dat2save} method
to save the relevant information in an object cleanly. It also has an
\code{explo.plot} method to plot the data contained in the object.
}

\author{Sonia Tarazona}

\keyword{classes}
