\name{Data2Save}
\alias{dat2save}

\docType{methods}

\title{
Saving data generated for exploratory plots.
}

\description{
This function is to save the data generated to draw the exploratory plots in a user-friendly format.
}

\value{
The dat2save() function takes the object generated by dat() and creates
a new one with the most relevant information.
}


\author{
Sonia Tarazona
}


\seealso{
       \code{\link{readData}}, \code{\link{addData}}, \code{\link{dat}},
       \code{\link{explo.plot}}.
     }

\examples{
## Load the input object with the expression data and the annotations
data(myCounts)

## Generating data for the plot "biodetection" and samples in columns 3 and 4 of expression data
mydata2plot = dat(mydata, type = "biodetection", k = 0)

## Save the relevant information cleanly
mydata2save = dat2save(mydata2plot)

}
