\name{Exploratory_Plots}
\alias{explo.plot}

\docType{methods}

\title{
Exploratory plots for sequencing data.
}

\description{
Different types of plots showing the biological classification for detected features, the comparison of count distributions, the distribution of counts or the length for detected features, the saturation over different sequencing depths, etc.
}

\value{
The explo.plot() function takes the object generated by dat() and draws the corresponding plot.
}


\author{
Sonia Tarazona
}


\seealso{
       \code{\link{readData}}, \code{\link{addData}}, \code{\link{dat}}.
     }

\examples{
## Load the input object with the expression data and the annotations
data(myCounts)

## Generating data for the plot "biodetection" and samples in columns 3 and 4 of expression data
mydata2plot = dat(mydata, type = "biodetection", k = 0, selection = c(3,4))

## Generating the corresponding plot
explo.plot(mydata2plot)

}
